core/gpu_hw_opengl: drop redundant outer static_cast in ConvertToFramebufferCoordinates
The expression was

  static_cast<s32>(static_cast<s32>(VRAM_HEIGHT) - y)

where VRAM_HEIGHT is constexpr u32 = 512 and y is s32. The inner
static_cast<s32>(VRAM_HEIGHT) is meaningful - it forces the
subtraction to s32 arithmetic (without it, u32 - s32 would promote
both operands to u32 via the usual arithmetic conversions, then
the make_tuple<s32, s32> would need a narrowing cast back). The
outer static_cast<s32> is a no-op since the subtraction already
produces s32.

Drop the outer cast. Behaviour and codegen unchanged.

+++ src/core/gpu_hw_opengl.cpp
@@ -847,7 +847,7 @@ void GPU_HW_OpenGL::UnmapBatchVertexPointer(u32 used_vertices)
 
 std::tuple<s32, s32> GPU_HW_OpenGL::ConvertToFramebufferCoordinates(s32 x, s32 y)
 {
-  return std::make_tuple(x, static_cast<s32>(static_cast<s32>(VRAM_HEIGHT) - y));
+  return std::make_tuple(x, static_cast<s32>(VRAM_HEIGHT) - y);
 }
 
 void GPU_HW_OpenGL::SetCapabilities()
